+++ modules/Scheduling/MassDrops.php
@@ -183,11 +183,11 @@ if ( $_REQUEST['modfunc'] != 'choose_course' )
 			echo $course_title . '<br />' . $period_title . '<br /><br />';
 		}
 
-		$popup_url = URLEscape( 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=choose_course' );
+		$popup_url = 'Modules.php?modname=' . $_REQUEST['modname'] . '&modfunc=choose_course';
 
-		echo '</div><a href="#" onclick="' . AttrEscape( 'popups.open(
-			' . json_encode( $popup_url ) . '
-			); return false;' ) . '">' . _( 'Choose a Course' ) . '</a></td></tr>';
+		// @since 12.0 Use colorBox instead of popup window
+		echo '</div><a href="' . URLEscape( $popup_url ) . '" class="colorbox">' .
+			_( 'Choose a Course' ) . '</a></td></tr>';
 
 		echo '<tr><td><br />' . DateInput(
 			DBDate(),
@@ -279,7 +279,13 @@ if ( $_REQUEST['modfunc'] === 'choose_course' )
 			FROM course_periods
 			WHERE COURSE_PERIOD_ID='" . (int) $_SESSION['MassDrops.php']['course_period_id'] . "'" );
 
-		echo '<script>opener.document.getElementById("course_div").innerHTML = ' .
-			json_encode( $course_title . '<br />' . $period_title ) . '; window.close();</script>';
+		// @since 12.0 Use colorBox instead of popup window
+		?>
+		<script>
+			document.getElementById("course_div").innerHTML = <?php echo json_encode( $course_title . '<br />' . $period_title ); ?>;
+
+			$.colorbox.close();
+		</script>
+		<?php
 	}
 }